final
i slowed down the parts that i saw had issues with the song timing. from
what i see this matches with what i was trying to accomplish, i hope it
works. the whole thing should probably be 30 seconds , if my first part
was 16s which i just extended that a bit to match the second part better
so the second part should be 14 seconds and then after 1:11 there is
silence so i have one pattern to hit at that time and hold for a while
to show it is the end.